# src/main/resources/static/swagger/index.yaml
swagger: '2.0'
info:
  description: cmp是一个客户关系管理平台
  version: 1.0.0
  title: 客户关系管理平台
basePath: /cam
schemes:
  - http

#认证
securityDefinitions:
  api_key:
    type: "apiKey"
    name: "token"
    in: "header"

#定义接口数据
paths:
...
  /mma/crmmediatemplate/list:
    get:
      tags:
        - 多媒体模板
      summary: 模板查询列表
      produces:
        - application/json
      parameters:
        - name: page
          description: 页码
          in: query
          type: integer
          required: true
        - name: limit
          description: 每页条数
          in: query
          type: integer
          required: true
        - name: templateName
          description: 模板名称
          in: query
          type: string
        - name: templateType
          description: 模板类型 1:短信 2:邮件
          in: query
          type: string
      responses:
        '200':
          description: 返回模板查询列表
          schema:
            $ref: '#/definitions/CrmMediaTemplateEntity'
  /mma/crmmediatemplate/info/{id}:
    get:
      tags:
        - 多媒体模板
      summary: 获取模板列表单条数据
      produces:
        - application/json
      parameters:
        - name: id
          description: ID
          in: path
          type: string
          required: true
      responses:
        '200':
          description: 返回模板列表单条数据
          schema:
            type: object
            properties:
              code:
                description: 状态码 0：成功  非0：失败
                type: integer
                format: int32
              user:
                $ref: '#/definitions/CrmMediaTemplateEntity'
  /mma/crmmediatemplate/save:
    post:
      tags:
        - 多媒体模板
      summary: 添加模板列表单条数据
      produces:
        - application/json
      parameters:
        - name: body
          description: 过滤器对象
          in: body
          type: string
          schema:
            $ref: '#/definitions/CrmMediaTemplateEntity'
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /mma/crmmediatemplate/update:
    post:
      tags:
        - 多媒体模板
      summary: 修改模板列表单条数据
      produces:
        - application/json
      parameters:
        - name: body
          description: 模板列表单条数据
          in: body
          type: string
          schema:
            $ref: '#/definitions/CrmMediaTemplateEntity'
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /mma/crmmediatemplate/delete:
    post:
      tags:
        - 多媒体模板
      summary: 删除模板颜色单条数据
      produces:
        - application/json
      parameters:
        - name: body
          description: 模板颜色id
          in: body
          type: array
          items:
            type: string
            default: 0
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'



  #...................多媒体模板附件表........................
  /mma/crmmediatemplatefile/list:
    get:
      tags:
        - 多媒体模板附件
      summary: 多媒体模板附件表查询
      produces:
        - application/json
      parameters:
        - name: page
          description: 页码
          in: query
          type: integer
          required: true
        - name: limit
          description: 每页条数
          in: query
          type: integer
          required: true
      responses:
        '200':
          description: 返回多媒体模板附件表
          schema:
            $ref: '#/definitions/CrmMediaTemplateFileEntity'
  /mma/crmmediatemplatefile/info/{id}:
    get:
      tags:
        - 多媒体模板附件
      summary: 获取多媒体模板附件表
      produces:
        - application/json
      parameters:
        - name: id
          description: ID
          in: path
          type: string
          required: true
      responses:
        '200':
          description: 返回多媒体模板附件表
          schema:
            type: object
            properties:
              code:
                description: 状态码 0：成功  非0：失败
                type: integer
                format: int32
              user:
                $ref: '#/definitions/CrmMediaTemplateFileEntity'
  /mma/crmmediatemplatefile/save:
    post:
      tags:
        - 多媒体模板附件
      summary: 添加多媒体模板附件表
      produces:
        - application/json
      parameters:
        - name: body
          description: 过滤器对象
          in: body
          type: string
          schema:
            $ref: '#/definitions/CrmMediaTemplateFileEntity'
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /mma/crmmediatemplatefile/update:
    post:
      tags:
        - 多媒体模板附件
      summary: 修改多媒体模板附件表
      produces:
        - application/json
      parameters:
        - name: body
          description: 多媒体模板附件表
          in: body
          type: string
          schema:
            $ref: '#/definitions/CrmMediaTemplateFileEntity'
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /mma/crmmediatemplatefile/delete:
    post:
      tags:
        - 多媒体模板附件
      summary: 删除多媒体模板附件表
      produces:
        - application/json
      parameters:
        - name: body
          description: 多媒体模板附件表id
          in: body
          type: array
          items:
            type: string
            default: 0
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'

  #...................多媒体变量表........................
...
  /mma/crmsmssendrecords/list:
    get:
      tags:
        - 短信发送记录
      summary: 短信发送记录表查询
      produces:
        - application/json
      parameters:
        - name: page
          description: 页码
          in: query
          type: integer
          required: true
        - name: limit
          description: 每页条数
          in: query
          type: integer
          required: true
        - name: customerName
          description: 客户姓名
          in: query
          type: string
        - name: isDelete
          description: 是否删除
          in: query
          type: string
        - name: templateType
          description: 模板类型
          in: query
          type: string
        - name: status
          description: 状态
          in: query
          type: string
        - name: sendNumber
          description: 客户号码
          in: query
          type: string
        - name: deliveryTime
          description: 定时发送时间
          in: query
          type: date
        - name: startTime
          description: 开始时间
          in: query
          type: date
        - name: endTime
          description: 结束时间
          in: query
          type: date
        - name: realSendTime
          description: 实际发送时间
          in: query
          type: string
        - name: batchFlag
          description: 发送批次标志
          in: query
          type: string
        - name: returnMsg
          description: 返回信息
          in: query
          type: string
      responses:
        '200':
          description: 返回短信发送记录表
          schema:
            $ref: '#/definitions/CrmSmsSendRecordsEntity'

  #...................短信发送记录表........................
  /mma/crmsmssendrecords/lists:
    post:
      tags:
        - 短信发送记录
      summary: 短信发送记录表查询1111
      produces:
        - application/json
      parameters:
        - name: body
          description: 过滤器对象
          in: body
          type: string
          schema:
            $ref: '#/definitions/smsFile'
          required: true
      responses:
        '200':
          description: 返回短信发送记录表
          schema:
            $ref: '#/definitions/smsFile'
  #...................短信接口........................
  /mma/crmsmssendrecords/sendMsgs:
    post:
      tags:
        - 短信发送记录
      summary: 调用第三方短信接口
      produces:
        - application/json
      parameters:
        - name: body
          description: 过滤器对象
          in: body
          type: string
          schema:
            $ref: '#/definitions/sendMsg'
          required: true
      responses:
        '200':
          description: 返回短信发送记录表
          schema:
            $ref: '#/definitions/returnMsg'
...
  /mma/crmsmssendrecords/info/{id}:
    get:
      tags:
        - 短信发送记录
      summary: 获取短信发送记录表
      produces:
        - application/json
      parameters:
        - name: id
          description: ID
          in: path
          type: string
          required: true
      responses:
        '200':
          description: 返回短信发送记录表
          schema:
            type: object
            properties:
              code:
                description: 状态码 0：成功  非0：失败
                type: integer
                format: int32
              user:
                $ref: '#/definitions/CrmSmsSendRecordsEntity'
  /mma/crmsmssendrecords/save:
    post:
      tags:
        - 短信发送记录
      summary: 添加短信发送记录表
      produces:
        - application/json
      parameters:
        - name: body
          description: 过滤器对象
          in: body
          type: string
          schema:
            $ref: '#/definitions/CrmSmsSendRecordsEntity'
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /mma/crmsmssendrecords/update:
    post:
      tags:
        - 短信发送记录
      summary: 修改短信发送记录表
      produces:
        - application/json
      parameters:
        - name: body
          description: 短信发送记录表
          in: body
          type: string
          schema:
            $ref: '#/definitions/CrmSmsSendRecordsEntity'
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /mma/crmsmssendrecords/delete:
    post:
      tags:
        - 短信发送记录
      summary: 删除短信发送记录表
      produces:
        - application/json
      parameters:
        - name: body
          description: 客户信息ID列表
          in: body
          type: array
          items:
            type: string
            default: 0
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'

  #...................短信发送记录表........................
...
definitions:
  R:
      type: object
      properties:
        code:
          description: 状态码 0：成功  非0：失败
          type: integer
          format: int32
        msg:
          description: 失败原因
          type: string
...
  smsFile:
    type: object
    properties:
      surname:
        description: 客户姓
        type: string
      customerName:
        description: 客户名
        type: string
      deliveryTime:
        description: 发送时间
        type: string
      mobile:
        description: 客户号码
        type: string
      status:
        description: 状态（已发送，发送失败，暂存）
        type: string
      templateType:
        description: 模板类型
        type: string
      startTime:
        description: 发送时间
        type: string
      endTime:
        description: 发送时间
        type: string
  #..............................短信发送接口........................
  sendMsg:
    type: object
    properties:
      command:
        description: 传固定值sendSMSMD5
        type: string
      username:
        description: 短信发送账号
        type: string
      pwd:
        description: 登录密码的MD5(32位加密且大写)
        type: string
      sendNumber:
        description: 手机号码（多个号码用英文逗号或者分号隔开），不能超过100个号码
        type: string
      content:
        description: 短信内容(最多140字符)
        type: string
      incode:
        description: 字符在传递过程中的编码,与outcode保持一致,GBK或UTF-8
        type: string
      outcode:
        description: 字符初始编码与incode保持一致, GBK或UTF-8
        type: string
      rstype:
        description: 可选参数
        type: string
      pfex:
        description: pfex为违禁词过滤开关 0表示开启 ; 1表示关闭 ; 空表示默认开启
        type: string
      extCode:
        description: 扩展码参数只允许数字字符
        type: string
  #..............................短信接收接口........................
  returnMsg:
    type: object
    properties:
      flag:
        description: 返回标示位
        type: string
      retMsg:
        description: 返回的信息
        type: string
        format: int32
...
  CrmMediaTemplateEntity:
    type: object
    properties:
      id:
        description: ID
        type: string
      templateName:
        description: 模板名称
        type: string
      templateType:
        description: 模板类型 1:短信 2:邮件
        type: string
      isEnabled:
        description: 是否启用 0:否 1:是
        type: string
      templateContent:
        description: 模板正文
        type: string
      isFile:
        description: 是否有附件 0:否 1:是
        type: string
      remark:
        description: 备注
        type: string
      createCode:
        description: 创建人代码
        type: string
      createName:
        description: 创建人
        type: string
      createTime:
        description: 创建时间
        type: string
      updateCode:
        description: 修改人代码
        type: string
      updateName:
        description: 修改人
        type: string
      updateTime:
        description: 修改时间
        type: string


  #........................多媒体模板附件表..................
  CrmMediaTemplateFileEntity:
    type: object
    properties:
      id:
        description: ID
        type: string
      templateId:
        description: 模板id
        type: string
      fileName:
        description: 文件名称
        type: string
      fileUrl:
        description: 文件名称地址
        type: string

  #........................多媒体变量表..................
...
  CrmSmsSendRecordsEntity:
    type: object
    properties:
      id:
        description: ID
        type: string
      templateId:
        description: 模板id
        type: string
      customerId:
        description: 客户id
        type: long
        format: int32
      sendNumber:
        description: 发送号码
        type: string
      content:
        description: 发送内容
        type: string
      status:
        description: 状态 1:成功 2:失败
        type: string
      isDelivery:
        description: 是否定时发送 0:否 1:是
        type: string
      deliveryTime:
        description: 定时发送时间
        type: string
      createCode:
        description: 创建人代码
        type: string
      createName:
        description: 创建人
        type: string
      createTime:
        description: 创建时间
        type: string
      realSendTime:
        description: 实际发送时间
        type: string
      batchFlag:
        description: 发送批次标志
        type: string
      returnMsg:
        description: 返回信息
        type: string
...